Move user list to a component
